requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
motor>=3.3.0
pymongo>=4.5.0
pydantic>=2.6.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
bcrypt>=4.0.1
python-dotenv>=1.0.0
email-validator>=2.1.0 